requirements.txt
streamlit
pandas
matplotlib
wordcloud
pillow
